+++ oapi-sdk/src/main/java/tech/blacklake/dev/openapi/sdk/event/model/mfg/MfgResultNotifyProgressReportAdjustDTO.java
@@ -2,6 +2,7 @@ package tech.blacklake.dev.openapi.sdk.event.model.mfg;
 
 import tech.blacklake.dev.openapi.sdk.event.model.BaseCallbackDTO;
 
+import java.math.BigDecimal;
 import java.util.List;
 
 public class MfgResultNotifyProgressReportAdjustDTO extends BaseCallbackDTO {
@@ -70,7 +71,7 @@ public class MfgResultNotifyProgressReportAdjustDTO extends BaseCallbackDTO {
 
         private Long progressReportRecordId;
 
-        private Integer reportAmount;
+        private BigDecimal reportAmount;
 
         private Long reportUnitId;
 
@@ -126,11 +127,11 @@ public class MfgResultNotifyProgressReportAdjustDTO extends BaseCallbackDTO {
             this.progressReportRecordId = progressReportRecordId;
         }
 
-        public Integer getReportAmount() {
+        public BigDecimal getReportAmount() {
             return reportAmount;
         }
 
-        public void setReportAmount(Integer reportAmount) {
+        public void setReportAmount(BigDecimal reportAmount) {
             this.reportAmount = reportAmount;
         }
